chore: keyboard shortcuts broken
