+++ app/pages/home/home.tsx
@@ -1,21 +1,16 @@
-import React, { FC, useRef, useCallback } from 'react';
+import React, { FC } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Spinner } from '@blockstack/ui';
-import { useHotkeys } from 'react-hotkeys-hook';
 
 import { Api } from '@api/api';
-import { increment, decrement } from '@utils/mutate-numbers';
-import { RootState } from '@store/index';
 import { openTxInExplorer } from '@utils/external-links';
+
+import { RootState } from '@store/index';
 import { selectAddress } from '@store/keys';
 import { selectActiveNodeApi } from '@store/stacks-node';
 import { selectAddressBalance, selectAvailableBalance } from '@store/address';
-import {
-  selectTransactionList,
-  selectTransactionsLoading,
-  selectTransactionListFetchError,
-} from '@store/transaction';
-import { selectPendingTransactions } from '@store/pending-transaction';
+import { selectRevokeDelegationModalOpen } from '@store/home/home.reducer';
+import { selectTransactionsLoading, selectTransactionListFetchError } from '@store/transaction';
 import { selectLoadingStacking, selectNextCycleInfo, selectStackerInfo } from '@store/stacking';
 import {
   homeActions,
@@ -24,6 +19,20 @@ import {
   selectHomeCardState,
   HomeCardState,
 } from '@store/home';
+
+import { TransactionModal } from '@modals/transaction/transaction-modal';
+import { ReceiveStxModal } from '@modals/receive-stx/receive-stx-modal';
+import { RevokeDelegationModal } from '@modals/revoke-delegation/revoke-delegation-modal';
+
+import { useDelegationStatus } from '@hooks/use-delegation-status';
+import { useTransactionList } from '@hooks/use-transaction-list';
+
+import { StackingCard } from '@components/home/stacking-card';
+import { StackingLoading } from '@components/home/stacking-loading';
+import { StackingBeginsSoonCard } from '@components/home/stacking-begins-soon-card';
+import { StackingError } from '@components/home/stacking-error-card';
+import { TransactionListItemMempool } from '@components/home/transaction-list/transaction-list-item-mempool';
+import { DelegationCard } from '@components/home/delegation-card';
 import {
   TransactionList,
   StackingPromoCard,
@@ -31,21 +40,7 @@ import {
   TransactionListItem,
   BalanceCard,
 } from '@components/home';
-import { TransactionModal } from '@modals/transaction/transaction-modal';
-import { ReceiveStxModal } from '@modals/receive-stx/receive-stx-modal';
-
-import { StackingCard } from '@components/home/stacking-card';
-import { StackingLoading } from '@components/home/stacking-loading';
-import { StackingBeginsSoonCard } from '@components/home/stacking-begins-soon-card';
-import { StackingError } from '@components/home/stacking-error-card';
-
 import { HomeLayout } from './home-layout';
-import { TransactionListItemMempool } from '@components/home/transaction-list/transaction-list-item-mempool';
-import { useMempool } from '@hooks/use-mempool';
-import { DelegationCard } from '@components/home/delegation-card';
-import { useDelegationStatus } from '@hooks/use-delegation-status';
-import { RevokeDelegationModal } from '@modals/revoke-delegation/revoke-delegation-modal';
-import { selectRevokeDelegationModalOpen } from '../../store/home/home.reducer';
 
 export const Home: FC = () => {
   const dispatch = useDispatch();
@@ -56,8 +51,6 @@ export const Home: FC = () => {
     address,
     balance,
     spendableBalance,
-    txs,
-    pendingTxs,
     loadingTxs,
     txModalOpen,
     txListFetchError,
@@ -68,9 +61,7 @@ export const Home: FC = () => {
     stackingCardState,
   } = useSelector((state: RootState) => ({
     address: selectAddress(state),
-    txs: selectTransactionList(state),
     spendableBalance: selectAvailableBalance(state),
-    pendingTxs: selectPendingTransactions(state),
     balance: selectAddressBalance(state),
     txModalOpen: selectTxModalOpen(state),
     revokeDelegationModalOpen: selectRevokeDelegationModalOpen(state),
@@ -84,37 +75,9 @@ export const Home: FC = () => {
     stackingCardState: selectHomeCardState(state),
   }));
 
-  const { mempoolTxs } = useMempool();
-
-  const focusedTxIdRef = useRef<string | null>(null);
-  const txDomNodeRefMap = useRef<Record<string, HTMLButtonElement>>({});
-
-  const focusTxDomNode = useCallback(
-    (shift: (i: number) => number) => {
-      const allTxs = [...mempoolTxs, ...pendingTxs, ...txs];
-      if (allTxs.length === 0) return;
-      if (focusedTxIdRef.current === null) {
-        const txId = allTxs[0].tx_id;
-        focusedTxIdRef.current = txId;
-        txDomNodeRefMap.current[txId].focus();
-        return;
-      }
-      const nextIndex = shift(allTxs.findIndex(tx => tx.tx_id === focusedTxIdRef.current));
-      const nextTx = allTxs[nextIndex];
-      if (!nextTx) return;
-      const domNode = txDomNodeRefMap.current[nextTx.tx_id];
-      if (!domNode) return;
-      domNode.focus();
-    },
-    [txs, pendingTxs, mempoolTxs]
-  );
-
-  useHotkeys('j', () => focusTxDomNode(increment), [txs, pendingTxs, mempoolTxs]);
-  useHotkeys('k', () => focusTxDomNode(decrement), [txs, pendingTxs, mempoolTxs]);
+  const { txs, pendingTxs, txCount, focusedTxIdRef, txDomNodeRefMap } = useTransactionList();
 
   if (!address) return <Spinner />;
-
-  const txCount = txs.length + pendingTxs.length + mempoolTxs.length;
 
   const transactionList = (
     <>
@@ -124,31 +87,16 @@ export const Home: FC = () => {
         node={activeNode}
         error={txListFetchError}
       >
-        {mempoolTxs
-          .filter(mempoolTx => !txs.map(tx => tx.tx_id).includes(mempoolTx.tx_id))
-          .map(mempoolTx => (
-            <TransactionListItemMempool
-              address={address}
-              domNodeMapRef={txDomNodeRefMap}
-              activeTxIdRef={focusedTxIdRef}
-              key={mempoolTx.tx_id}
-              tx={mempoolTx}
-              onSelectTx={openTxInExplorer}
-            />
-          ))}
-        {pendingTxs
-          .filter(pendingTx => !txs.map(tx => tx.tx_id).includes(pendingTx.tx_id))
-          .filter(pendingTx => !mempoolTxs.map(tx => tx.tx_id).includes(pendingTx.tx_id))
-          .map(pendingTx => (
-            <TransactionListItemMempool
-              address={address}
-              domNodeMapRef={txDomNodeRefMap}
-              activeTxIdRef={focusedTxIdRef}
-              key={pendingTx.tx_id}
-              tx={pendingTx}
-              onSelectTx={openTxInExplorer}
-            />
-          ))}
+        {pendingTxs.map(pendingTxs => (
+          <TransactionListItemMempool
+            address={address}
+            domNodeMapRef={txDomNodeRefMap}
+            activeTxIdRef={focusedTxIdRef}
+            key={pendingTxs.tx_id}
+            tx={pendingTxs}
+            onSelectTx={openTxInExplorer}
+          />
+        ))}
         {txs.map(tx => (
           <TransactionListItem
             domNodeMapRef={txDomNodeRefMap}
